adds frameRate function sets frameRate to 60. need to ckeck out examples for speed of rendering

+++ src/Graphics/Proc/Core/State/Elements/Frame.hs
@@ -8,7 +8,7 @@ import Data.Default
 
 data FrameState = FrameState 
   { frameCount :: Int
-  , frameRate  :: Int
+  , frameRate  :: Float
   , frameLoop  :: LoopMode
   }
